Forecast wrapper V2 + demo
We wanted a way to save both daily and hourly weather data, but with
the old wrapper you needed 2 calls to Forecast to do this. This update
fixes that by using a single object `Weather` which holds all data and
methods `days()` and `hours()` to access them.
Some arguments were also renamed to be more verbose and pythonic, so be
sure to check out the new demo.