got rid of explicit column names
It used to be that you needed to pass a single dataframe, and the names
of the columns that the analysis should use, but imho this is stupid: i
just want to pass two series: the independent variable and the
dependent variable. This however breaks our ‘black box’-model a little
bit… @saroele what do you think?