selectable pickle format
We ran into some problems saving and reloading a custom made houseprint
object using jsonpickle. I was able to fix it using the normal pickle,
so I’ve added an argument `pickle_format`, which allows you to choose
which format you want to save your houseprints to. @saroele I have done
this so that everything is backwards compatible, but I believe that we
should just go back to the normal pickle, since we are pushing everyone
to python3.